requirements.txt
django==5.1.7
python-dotenv==1.1.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.164.0
